Refactor: 코드 구조 개선으로 인한, 매개변수 및 파일 위치 변경
