--- backend/server/src/main/java/com/todoc/server/common/exception/CustomException.java
@@ -1,15 +1,0 @@
-package com.todoc.server.common.exception;
-
-public abstract class CustomException extends RuntimeException{
-
-    private final ErrorCode errorCode;
-
-    public CustomException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-
-    public ErrorCode getErrorCode() {
-        return errorCode;
-    }
-}
