--- backend/server/src/main/java/com/todoc/server/common/exception/ErrorCode.java
@@ -1,9 +1,0 @@
-package com.todoc.server.common.exception;
-
-import org.springframework.http.HttpStatus;
-
-public interface ErrorCode {
-
-    HttpStatus getStatus();
-    String getMessage();
-}
